core/standards/scripts/onload/064.html

<!DOCTYPE html>
<html>
  <title>Removing slow-loading rel="stylesheet" with slow-loading subresource doesn't stop load delay</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>
  <div id="test">Test</div>
  <link rel="stylesheet" href="_files/background.css">
  <script>
    var t = async_test(undefined, {timeout:4000});
    events = [];
    onload = function() {
      events.push("load");
      setTimeout(t.step_func(function() {
        document.body.style.width = ""; 
        assert_array_equals(events, ["load", "error"]);
        t.done()
    }), 500)}
    addEventListener("DOMContentLoaded",
      function() {setTimeout(t.step_func(function() {

                  var link = document.getElementsByTagName("link")[0];
    
                  document.body.style.width = "0";

                  link.parentNode.removeChild(link);

                  var img2 = document.createElement("img");
                  img2.onerror = function() {events.push("error")};
                  img2.src = "#abc"
                  document.body.appendChild(img2);
      }), 200)}, false);
  </script>
